requirements.txt
ctransformers==0.2.27
sentence-transformers==3.0.1
langchain==0.2.6
flask
pypdf==4.2.0
faiss-cpu==1.8.0
streamlit==1.18.0
streamlit-chat==0.1.1
altair==4.2.2
protobuf==3.19.0
-e .
